parser: abstract call argument parsing logic
